requirements.txt
cyclonedx-python-lib==3.1.5